# workers/swagger.json
{"openapi":"3.1.0","info":{"title":"API","version":"0.1.0"},"paths":{"/api/tts/index-tts":{"post":{"tags":["TTS"],"summary":"Submit IndexTTS task","description":"提交IndexTTS任务。`params` 参数会直接传给 `IndexTTS2.infer()`。\n\n**可传字段示例**：\n\n- `text`: str, 要合成的文本\n- `spk_audio_prompt`: str, 参考音频路径，用于克隆说话人\n- `emo_vector`: list[float], 8维情感向量 `[happy, angry, sad, afraid, disgusted, melancholic, surprised, calm]`\n- `cfg_value`: float, LM引导强度\n- `inference_timesteps`: int, 推理步数\n- `normalize`: bool, 是否启用外部文本归一化\n- `denoise`: bool, 是否启用降噪\n- `retry_badcase`: bool, 是否重试坏案例\n- `retry_badcase_max_times`: int, 最大重试次数\n- `retry_badcase_ratio_threshold`: float, 坏案例检测长度阈值\n- `use_random`: bool, 是否启用随机性\n- `output_path`: str, 输出wav路径\n- `verbose`: bool, 是否打印详细推理信息\n\n**示例 JSON**：\n\n```json\n{\n  \"params\": {\n    \"text\": \"hello\",\n    \"spk_audio_prompt\": \"examples/voice_01.wav\",\n    \"emo_vector\": [0,0,0,0,0,0,0,0],\n    \"cfg_value\": 2,\n    \"inference_timesteps\": 10,\n    \"normalize\": false,\n    \"denoise\": false,\n    \"retry_badcase\": true,\n    \"retry_badcase_max_times\": 3,\n    \"retry_badcase_ratio_threshold\": 6,\n    \"use_random\": false,\n    \"output_path\": \"tasks/output.wav\",\n    \"verbose\": true\n  }\n}","operationId":"submit_index_tts_api_tts_index_tts_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TTSRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/tts/voxcpm":{"post":{"tags":["TTS"],"summary":"Submit VoxCPM task","description":"提交VoxCPM任务。`params` 参数会直接传给 `VoxCPM.generate()`。\n\n**可传字段示例**：\n\n- `text`: str, 要合成的文本\n- `prompt_wav_path`: str, 参考音频路径，用于克隆说话人 (对应 spk_audio_prompt)\n- `prompt_text`: str, 参考音频对应文本\n- `cfg_value`: float, guidance 强度（1.5~2.5 推荐）\n- `inference_timesteps`: int, LocDiT 推理步数（越大越慢但更稳）\n- `normalize`: bool, 是否保留原始文本（推荐 False）\n- `denoise`: bool, 是否启用 16kHz 降噪（推荐 False）\n- `retry_badcase`: bool, 是否启用异常自动重试\n- `retry_badcase_max_times`: int, 最多重试次数\n- `retry_badcase_ratio_threshold`: float, 语音过长判定阈值\n\n**示例 JSON**：\n\n```json\n{\n  \"params\": {\n    \"text\": \"Translate for me, what is a surprise!\",\n    \"prompt_wav_path\": \"examples/voice_01.wav\",\n    \"prompt_text\": \"Translate for me, what is a surprise!\",\n    \"cfg_value\": 2.0,\n    \"inference_timesteps\": 10,\n    \"normalize\": false,\n    \"denoise\": false,\n    \"retry_badcase\": true,\n    \"retry_badcase_max_times\": 3,\n    \"retry_badcase_ratio_threshold\": 6.0\n  }\n}","operationId":"submit_voxcpm_tts_api_tts_voxcpm_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TTSRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/tts/{task_id}":{"get":{"tags":["Tasks"],"summary":"Get a specific TTS task by ID","description":"Get details of a specific TTS task by its ID.\n\n**Path Parameters**:\n\n- `task_id`: str, The unique identifier of the task (UUID).\n\n**Response**:\n\nA task object containing:\n- `id`: str, Unique task identifier (UUID).\n- `status`: str, Current task status (`pending`, `processing`, `completed`, `failed`).\n- `result`: str or null, Result output (e.g., \"processed\" for completed tasks).\n- `audio_data`: str or null, Base64-encoded audio data with data URI prefix (only present if status is `completed` and audio file exists).\n- `worker_id`: str or null, ID of the worker processing the task.\n- `submit_time`: float, Timestamp when the task was submitted (from created_at).\n- `start_time`: float or null, Timestamp when processing started (not directly tracked).\n- `end_time`: float or null, Timestamp when processing ended (from updated_at).\n- `duration`: float or null, Processing time in seconds.\n- `params`: dict, TTS parameters used for the task.\n- `error_msg`: str or null, Error message (only present if status is `failed`).\n- `retry_time`: int, Remaining retry attempts (default 3).\n- `task_result`: dict or null, Result data from processing.\n- `task_type`: str or null, Type of the task (e.g., \"index-tts\", \"voxcpm\").\n\n**Example Response** (status: completed):\n\n```json\n{\n  \"id\": \"123e4567-e89b-12d3-a456-426614174000\",\n  \"status\": \"completed\",\n  \"result\": \"processed\",\n  \"task_result\": {\"output\": \"processed\", \"duration\": 1.23},\n  \"worker_id\": \"worker-uuid\",\n  \"submit_time\": 1700000000.0,\n  \"duration\": 1.23,\n  \"params\": {\"text\": \"Hello\"},\n  \"task_type\": \"index-tts\"\n}\n```","operationId":"get_task_api_tts__task_id__get","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/workers":{"get":{"tags":["Internal"],"summary":"Get all registered workers","description":"Returns a list of all workers currently registered in the system, including their platform, memory, CPU, and GPU information.","operationId":"get_workers_api_workers_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WorkersResponse"}}}}}}},"/api/tasks":{"get":{"tags":["Tasks"],"summary":"Query tasks","description":"Retrieve tasks with optional filters for task_type, status, and limit. Use status=all or omit status to get all statuses. Use all=true to get all tasks without limit.","operationId":"get_tasks_api_tasks_get","parameters":[{"name":"task_type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by task type","title":"Task Type"},"description":"Filter by task type"},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by status (use 'all' or omit for all statuses)","title":"Status"},"description":"Filter by status (use 'all' or omit for all statuses)"},{"name":"all","in":"query","required":false,"schema":{"type":"boolean","description":"If true, return all tasks without limit","default":false,"title":"All"},"description":"If true, return all tasks without limit"},{"name":"limit","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"description":"Maximum number of tasks to return","default":10,"title":"Limit"},"description":"Maximum number of tasks to return"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TasksResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/upload":{"post":{"tags":["Files"],"summary":"Upload a file","description":"Upload a file to the media directory.\n\n- **file**: The file to upload (binary data).\n- **rel_path**: Optional relative path within the media directory (e.g., \"images/\"). Must not contain \"..\" for security.\n\nThe file is saved to `{media_path}/{rel_path}/{filename}`. Returns a download URL.","operationId":"upload_file_api_upload_post","parameters":[{"name":"rel_path","in":"query","required":false,"schema":{"type":"string","default":"","title":"Rel Path"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_file_api_upload_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/file":{"get":{"tags":["Files"],"summary":"Download a file","description":"Download a file from the media directory.\n\n- **path**: The relative path to the file within the media directory (e.g., \"images/photo.jpg\"). Must not contain \"..\" for security.\n\nReturns the file as a binary response.","operationId":"get_file_api_file_get","parameters":[{"name":"path","in":"query","required":true,"schema":{"type":"string","title":"Path"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/queue/length":{"get":{"tags":["Queue"],"summary":"Get Queue Length","operationId":"get_queue_length_api_queue_length_get","parameters":[{"name":"task_type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Task Type"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/next_task":{"get":{"tags":["Queue"],"summary":"Get next pending task","description":"Retrieves the next pending task from the queue, optionally filtered by task type. The task status is automatically updated to 'processing'.","operationId":"get_next_task_api_next_task_get","parameters":[{"name":"task_type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter tasks by type (e.g., 'test' for TTS, 'whisper' for transcription). If not specified, returns the next task of any type.","title":"Task Type"},"description":"Filter tasks by type (e.g., 'test' for TTS, 'whisper' for transcription). If not specified, returns the next task of any type."}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/whisper/audio/url":{"post":{"tags":["Whisper"],"summary":"Submit Whisper Audio Url","operationId":"submit_whisper_audio_url_api_whisper_audio_url_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WhisperRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/whisper/video/url":{"post":{"tags":["Whisper"],"summary":"Submit Whisper Video Url","operationId":"submit_whisper_video_url_api_whisper_video_url_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WhisperRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/whisper/audio/data":{"post":{"tags":["Whisper"],"summary":"Submit Whisper Audio Data","operationId":"submit_whisper_audio_data_api_whisper_audio_data_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_submit_whisper_audio_data_api_whisper_audio_data_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_submit_whisper_audio_data_api_whisper_audio_data_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_submit_whisper_audio_data_api_whisper_audio_data_post"},"Body_upload_file_api_upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_file_api_upload_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"TTSRequest":{"properties":{"params":{"additionalProperties":true,"type":"object","title":"Params","description":"Parameters for TTS inference, passed directly to the TTS model."}},"type":"object","required":["params"],"title":"TTSRequest","example":{"params":{"speed":1.0,"text":"Hello, world!","voice":"en-US-1"}}},"TasksResponse":{"properties":{"total":{"type":"integer","title":"Total"},"tasks":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Tasks"}},"type":"object","required":["total","tasks"],"title":"TasksResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WhisperRequest":{"properties":{"url":{"type":"string","title":"Url"}},"type":"object","required":["url"],"title":"WhisperRequest"},"WorkerModel":{"properties":{"id":{"type":"string","title":"Id","description":"Unique identifier for the worker"},"platform":{"type":"string","title":"Platform","description":"Operating system and platform information"},"memory_total":{"type":"integer","title":"Memory Total","description":"Total memory in bytes"},"memory_available":{"type":"integer","title":"Memory Available","description":"Available memory in bytes"},"cpu_count":{"type":"integer","title":"Cpu Count","description":"Number of CPU cores"},"cpu_freq":{"type":"number","title":"Cpu Freq","description":"CPU frequency in MHz"},"gpu_info":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Gpu Info","description":"GPU information if available"},"start_time":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Start Time","description":"Worker registration timestamp"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At","description":"Last update timestamp"}},"type":"object","required":["id","platform","memory_total","memory_available","cpu_count","cpu_freq"],"title":"WorkerModel","example":{"cpu_count":4,"cpu_freq":2200.0,"gpu_info":"NVIDIA GeForce GTX 1060","id":"worker-123","memory_available":4294967296,"memory_total":8589934592,"platform":"Linux-5.4.0-x86_64-with-glibc2.29"}},"WorkersResponse":{"properties":{"workers":{"items":{"$ref":"#/components/schemas/WorkerModel"},"type":"array","title":"Workers"}},"type":"object","required":["workers"],"title":"WorkersResponse"}}}}